requirements.txt
# Core data analysis
pandas>=2.0,<3.0
numpy>=1.24,<3.0

# PostgreSQL connectivity
psycopg2-binary>=2.9,<3.0

# Visualization
matplotlib>=3.7,<4.0

# Jupyter environment
jupyter>=1.0,<2.0
ipykernel>=6.25,<7.0

# Environment variables
python-dotenv>=1.0,<2.0